Add portable server fetch proxy helper
