+++ packages/developer-sdk/README.md
@@ -25,6 +25,26 @@ import { createSwigRouteHandlers } from '@swig-wallet/developer-sdk/server/next'
 export const { POST } = createSwigRouteHandlers();
 ```
 
+For Expo, React Native, or any app that does not host its own API routes, mount
+the same proxy on any Fetch-standard server:
+
+```typescript
+import { createSwigFetchHandler } from '@swig-wallet/developer-sdk/server/fetch';
+
+export default {
+  fetch: createSwigFetchHandler(),
+};
+```
+
+Point the mobile app at that deployed proxy:
+
+```typescript
+const swig = new SwigBrowserClient({
+  proxyUrl: 'https://api.example.com/swig',
+  network: 'devnet',
+});
+```
+
 Then the browser code can prepare transactions without knowing about that proxy:
 
 ```typescript
@@ -195,7 +215,8 @@ Set `SWIG_LOCAL_SMOKE_SUBMIT=false` to stop after preparing transactions without
 - `src/browser.ts` owns browser-safe wallet handles that prepare through an app proxy.
 - `src/core` owns HTTP transport, retry defaults, and SDK errors.
 - `src/passkeys` wraps Swig passkey signing helpers.
-- `src/server/next.ts` provides the Next.js catch-all proxy route helper.
+- `src/server/fetch.ts` provides the portable Fetch-standard proxy handler.
+- `src/server/next.ts` wraps the Fetch handler for Next.js catch-all routes.
 - `src/transactions` owns signed transaction submission, including sponsored send.
 - `src/types` contains the public TypeScript contracts split by concern.
 - `src/wallets` owns wallet handles, wallet operation clients, request shaping, and response normalization.
